cost-calculator/requirements.txt
attr==0.3.1
attrs==19.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2>=2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.0
py-dateutil==2.2
pycparser==2.19
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
